ci: add github actions for build and test
Squashed commit of the following:

commit 7b6e2ed618b34fc79a916af04d2d1d563d787fef
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:54:17 2023 +0800

    fix badge

commit c7130bd4f0a55672383c5cb74ff7bc3db65c902c
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:39:12 2023 +0800

    try fix install

commit 4ce3156ae6b2bce13ceee18088f4927f2119e58b
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:38:05 2023 +0800

    less github action warning by new ninja plugin

commit 694d7fd99293a9f7718f65bd3da4050ad71d402b
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:32:04 2023 +0800

    fix install for linux and macos

commit 2cac747427bf93a20936f1b0e5e762c6511609d4
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:20:25 2023 +0800

    specify ninja when configure instead of build

commit ddfcee735e0c6d8deaa9cc18b2d912751068baf2
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:18:43 2023 +0800

    Ninja as generator

commit 6e833cde1c07b29167b82956a8bb6c205d05435f
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:17:12 2023 +0800

    fix default install dir

commit 70b712c2c083f727dd3b2e5194410dd4ab43afdc
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:13:38 2023 +0800

    add badge for workflow status in README

commit bacaf488f8b80196568fcb60af99b82fa629cf83
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:13:00 2023 +0800

    run uninttests

commit 8d939271dcf2dadd2548ad4c40c6ff5cff2e3bd6
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Mon May 1 00:08:11 2023 +0800

    install the project for upload

commit bad469fda441b72dfbc56ff4d82068bbed34dc67
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:52:19 2023 +0800

    parallel build in CI

commit 834c7aa881f6972554c6ebbbca8b53b107d9592f
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:49:05 2023 +0800

    disable one unittest

commit e7a706a0f432255e76fc05557a0138c2073c66dc
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:34:34 2023 +0800

    treat C4244 as warning

commit a526c24aa5519f6fd0f038e0107bfd69ece03383
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:20:52 2023 +0800

    fix warnings on MSVC

commit 5de9fbccc3c3036ffe1fa927dc18451cfb7c8ccf
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:04:38 2023 +0800

    cache for deps

commit fb80ee49310817e4b0eb78ddb11ff776ec96976c
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 23:02:25 2023 +0800

    disable openexr

commit 7b44310d44748a4486aecc542e7903f000baed58
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:48:14 2023 +0800

    Xinerama for glfw

commit 89f0f9efb06ab0db5a9a5166ade89ff76d9bfed3
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:41:11 2023 +0800

    fix severe warnings reported by overlook

commit 5988c379ae3929cbd38d914d7b2d9994725ce76e
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:35:29 2023 +0800

    install xrandr

commit 4f57690166a00b88cc41a6b5184ed3c185244122
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:21:57 2023 +0800

    turn of mirror if in github action env

commit 64cf7256c8f06d43007dd44c501a4726cb460adf
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:05:46 2023 +0800

    python3 and pip3

commit 506ed6f9bf007336e85ff54554547722d47847e0
Author: zhuo@mbp <imzhuo@foxmail.com>
Date:   Sun Apr 30 22:01:03 2023 +0800

    fix opencv path on non-Windows

commit 2a7c934ca938c1c874360a4345937360863fc768
Author: Zhuo Zhang <imzhuo@foxmail.com>
Date:   Sun Apr 30 21:51:51 2023 +0800

    ci: build for windows, linux and mac x64

+++ src/kantu/transform_format.cpp
@@ -88,7 +88,7 @@ static void RGBfromYUV_BT601_u8(uint8_t& R, uint8_t& G, uint8_t& B, uint8_t Y, u
     static const int ITUR_BT_601_SHIFT = 20;
 
 
-    #define CLIP_TO_UCHAR(x) ((x) < 0 ? 0 : ((x) > 255 ? 255 : (x)))
+    #define CLIP_TO_UCHAR(x) (uint8_t)((x) < 0 ? 0 : ((x) > 255 ? 255 : (x)))
 
     int u = U - 128;///
     int v = V - 128;///
